example_sdk_client/requirements.txt
omotes-sdk-python==3.2.2
